Made skill tab almost fully functional
Made some very minor code changes (literally almost everything where IntelliJ gave a warning)
Rewrote lots of stuff in Skill
Added lots of new functionality to GuiButtonSkill
Made GuiSkillTab almost work (child skill buttons still need to be added)
Made silver tools and armor be repairable with silver (apparently it was removed when the new registry system was added)
Made SkillUnlockMessage work correctly with the new stuff in Skill
Added ClientHelper to have a Helper class for rendering and translation utilities
Added new number and collection utilities to GeneralHelper
Reordered lang file a bit and added all lang keys for the skills
Added dirt to skill window image because it's easier than to read through opengl documentation

+++ src/main/java/theblockbox/huntersdream/blocks/BlockWolfsbane.java
@@ -38,7 +38,7 @@ public class BlockWolfsbane extends BlockCrops {
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
-		return WOLFSBANE_AABB[state.getValue(this.getAgeProperty()).intValue()];
+		return WOLFSBANE_AABB[state.getValue(this.getAgeProperty())];
 	}
 
 	@Override
